backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
aiomysql>=0.2.0
redis>=5.0.1
celery>=5.3.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.1.0
python-multipart>=0.0.9
requests>=2.31.0
cryptography>=42.0.0
# Testing
pytest>=8.0.0
httpx>=0.26.0
